Prisma: get data only relevant to current platform
Previous behavior was to get all WalletKeys and Deposits from the
database. However, this could pose an issue in the situation where
multiple instances of the bot are running (e.g. multiple Chronik
WS subscriptions to the same Script causing the same deposit to be
saved to the database)

This commit changes the behavior of the Prisma queries to poll
only data associated with users of the current-running platform.