Load bot WalletKey on _initWalletManager()
`platform-fixes-1` introduced a bug where the WalletKey for the bot
would only be loaded on the first run; subsequent runs would skip
loading the bot key.

This commit adds the conditional for properly loading the bot key
on subsequent runs, and also adds methods to `database.ts` in order
to handle this new logic.