Properly check inputAmount of withdrawal tx
Previous behavior was using an evidently `undefined` property of the
bitcore-lib-xpi `Transaction` class. This was causing all available
UTXOs to be used in the withdrawal transaction when it may not have been
necessary. This commit uses the correct `Transaction` property in order
to add only required inputs to the withdrawal transaction.