Restructure WalletKey init and Chronik WS handling
This commit addresses a bug introduced in `bugfix-1`, where the `Script`
would not convert back into an `Address` for unsubscribing from Chronik
WS. It is fixed by adjusting the `WalletKey`s to store the Chronik
data (i.e. scriptHex, scriptType) instead of generating it on-the-fly.

Also use `_getScriptFromAddress` in some places, and move the Chronik
`AddedToMempool` event handler into `_chronikHandleWsMessage`.