Fix duplicate UTXOs to in-memory set
Behavior was observed where the Chronik WS would trigger duplicate
`AddedToMempool` events for the same UTXO. This was corrupting the
in-memory UTXO set. This commit adds a conditional to check if a UTXO
that triggers the `AddedToMempool` event is already present in our
in-memory set, and if so, ignores it.

This commit also fixes the extra `}` that was accidentally included in
the previous commit (the `}` was part of the conditional included in
this commit).