fix Client apiKey + clearInterval on shutdown
Previous commit introduced bug where the Client would not properly
authenticate with Discord without having `this.client.token` set with
the `apiKey`. This commit fixes that.

Another bug was noticed where the `setInterval()` for updating the bot
activity was not being cleared on shutdown, causing the node.js process
to hang until force-killed. This commit adds the `clearInterval()` to
the `stop()` method to resolve this.