Discord adjustments (#15)
* Many formatting changes, several optimizations

In summary:
- Convert many try/catch handlers to throw `Error`s rather than log
to console
- Adjust `setup()` and `launch()` logic; remove `this.rest` and use
`this.client.rest` instead so that we can pass `apiKey` without needing
to instantiate a private class property
- Move the `this.client.on('ready')` handler to a private class method
- Add new `_setRandomActivity()` method which is called during
`_handleReady()` to set the bot's current activity. Activities are
stored in a class property `activities: string[]` and the available
activity strings are pushed during `setup()`.
- Many more formatting fixes

* fix Client apiKey + clearInterval on shutdown

Previous commit introduced bug where the Client would not properly
authenticate with Discord without having `this.client.token` set with
the `apiKey`. This commit fixes that.

Another bug was noticed where the `setInterval()` for updating the bot
activity was not being cleared on shutdown, causing the node.js process
to hang until force-killed. This commit adds the `clearInterval()` to
the `stop()` method to resolve this.

---------

Co-authored-by: maff1989 <maff1989@gmail.com>