Convert "Give" command to on-chain transaction (#19)
This commit overhauls the LotusBot and WalletManager modules. When a
platform user issues the "Give" command, the transaction now occurs
on-chain vs. simply creating an entry in the database. Due to this,
all user balances are determined by the WalletManager.

The database Give schema has been updated to accommodate the txid of
a Give.

The platforms and the Platform interface were updated to receive the
txid from LotusBot after an on-chain give is successful. This allows
the platform to send the txid with its "give successful" message.

Several other tweaks are included in this commit, e.g. removing the
Bot account, adding a `_reconcileUtxos` method to WalletManager (for
ensuring user UTXOs are valid before balance is calculated), updated
logic for the Chronik WS handler, etc.

Co-authored-by: maff1989 <maff1989@gmail.com>