hotfix: accept utxo on `AddedToMempool`
v2.0.0 introduced a bug where user balance on Give would not be updated
until the utxo was confirmed. This commit allows the utxo to be updated
in the user's balance on `AddedToMempool` event.

This should be safe since we already implemented `_reconcileUtxos` in
WalletManager, which will remove missing/spent utxos from the user's
WalletKey and not include them in balance calculation.