Remove all tx confirmation logic (#23)
The confirmation logic was implemented so that users could not give or
withdraw Lotus without their deposit having first confirmed. Since
v2.0.0, Give interactions are now on-chain. This created a problem where
gives would trigger the Chronik `AddedToMempool` and be subject to the
same confirmation logic as deposits. However, this is terrible for the
expected UX, where gives are immediately re-givable.

Operating on 0-conf creates a potential double-spend DoS vector, where
a depositing user double-spends their deposit, thereby causing the bot
to retain an invalid UTXO in their set. However, this DoS vector is
mitigated with the `WalletManager._reconcileUtxos` method, which runs at
the same time `WalletManager.getUserBalanace` runs.

Co-authored-by: maff1989 <maff1989@gmail.com>