bugfix: give/withdraw race condition
In v2.0.0, a race condition was introduced where the Give/Withdrawal
would need to be saved to the database before the transaction triggered
the Chronik `AddedToMempool` event, otherwise the change from the
Give/Withdrawal transaction would be processed as a Deposit.

This commit adjusts the Give/Withdrawal transaction broadcasting logic
so that the Give/Withdrawal is first saved to the database before the
transaction is broadcasted. If the transaction fails to broadcast, then
the saved Give/Withdrawal is removed from the database.