bugfix: enforce withdraw to external wallet
A bug exists where a user could submit their deposit address as the
address for withdrawal. However, there is code that checks for and
ignores deposits with txids that exist in the Withdrawals database.
This commit will fail a withdrawal if the user submits their deposit
address in the `withdraw` command.