Optimize give/withdrawal tx processing
Previous behavior was to reconcile a user's utxos after broadcasting
the Give/Withdrawal. However, the user's utxos are already reconciled
when their balance is calculated. This commit adds a `usedUtxoCount`
to the `WalletManager._genTx`, which is used to splice the user's
utxo array after their Give/Withdrawal is broadcasted. It accomplishes
the same thing without having to reconcile utxos twice in one operation.