Moar bugfixes and tweaks (#24)
* bugfix: give/withdraw race condition

In v2.0.0, a race condition was introduced where the Give/Withdrawal
would need to be saved to the database before the transaction triggered
the Chronik `AddedToMempool` event, otherwise the change from the
Give/Withdrawal transaction would be processed as a Deposit.

This commit adjusts the Give/Withdrawal transaction broadcasting logic
so that the Give/Withdrawal is first saved to the database before the
transaction is broadcasted. If the transaction fails to broadcast, then
the saved Give/Withdrawal is removed from the database.

* LotusBot: Give/Withdrawal logic/formatting tweaks

* bugfix: round XPI->sats to nearest integer

Behavior was observed where converting an XPI unit to satoshis would
still result in a floating-point number.

Example: 0.010001 / XPI_DIVISOR = 1000.99999999...

This commit ensures the result of the conversion is rouned to the
nearest whole integer.

* bugfix: enforce withdraw to external wallet

A bug exists where a user could submit their deposit address as the
address for withdrawal. However, there is code that checks for and
ignores deposits with txids that exist in the Withdrawals database.
This commit will fail a withdrawal if the user submits their deposit
address in the `withdraw` command.

* Optimize give/withdrawal tx processing

Previous behavior was to reconcile a user's utxos after broadcasting
the Give/Withdrawal. However, the user's utxos are already reconciled
when their balance is calculated. This commit adds a `usedUtxoCount`
to the `WalletManager._genTx`, which is used to splice the user's
utxo array after their Give/Withdrawal is broadcasted. It accomplishes
the same thing without having to reconcile utxos twice in one operation.

* Logging/formatting tweaks