remove extra WalletKey utxo consolidation
Since we are doing a `_reconcileUtxos` when the user's balance is
calculated, there is no need to manually remove used utxos from the
WalletKeys after a tx is broadcasted. The next wallet action requested
by a user in LotusBot will check balance first, which will reconcile
the WalletKey's utxo set and remove the `SPENT` utxos.