Multi platform (#29)
* prepare systemd services for multi-platform

We will be running all platforms in a single node.js process, so there
is no longer a need to have a service file for each platform.

* prepare platforms/config/database

- Platform names in config.ts were lowercased
- Database logic adjusted to allow polling tables by specified platform
- Platform event emitters were adjusted to include the name of the
platform in the event

* Add multi-platform support + "link" preparations

This commit adds the multi-platform support to the LotusBot module. It
also sets up some LotusBot/WalletManager logic to prepare for adding the
"link" command, which will allow users to link their balance between
platforms.

* bugfix: get all database deposits when reconciling

Previous database call was not getting all deposits in the database;
this commit fixes that problem, so that we aren't throwing errors trying
to save duplicate deposits.

* undo LotusBot "link" preparation

we're not ready for this one yet :)

* remove extra WalletKey utxo consolidation

Since we are doing a `_reconcileUtxos` when the user's balance is
calculated, there is no need to manually remove used utxos from the
WalletKeys after a tx is broadcasted. The next wallet action requested
by a user in LotusBot will check balance first, which will reconcile
the WalletKey's utxo set and remove the `SPENT` utxos.