bugfix: premature return from Chronik WS handler (#30)
`_chronikHandleWsMessage` method was returning on emit, but this could
skip saving utxos for other local addresses that were involved in the tx
(i.e. a give tx usually adds 2 utxos). This commit removes the
premature return to ensure all utxos of a tx are processed accordingly.
Also, `break` from the loop after emitting, so that we don't
unnecessarily process user scriptHex against the same utxo outputScript.