Add link command (#33)
* implement "link" command

This commit includes all required functionality for the "link" command
to work on both Telegram and Discord platforms.

* bugfix: reverse WalletManager.updateKey params

Had these reversed, causing an undefined error when trying to update
a WalletKey on successful link. This commit fixes that method to
properly pass the data to the method.

* bugfix: create account first

A bug was found where a user that doesn't yet exist, and tries to run
the "link" command, will fail. This commit ensures the account exists
first before processing the command.

* bugfix: include sendLinkReply on twitter module

This should have been included in #31