bugfix: `_reconcileUtxos` removing valid utxos
Behavior was observed where a give following another give would result
in a Chronik tx broadcast error of:
`Bitcoind rejected tx: txn-mempool-conflict`

The cause is the `WalletManager._reconcileUtxos` method. If multiple
inputs were spent during the first give, and the total available utxos
was greater than 2, the following would happen:
1. first (spent) input utxo is spliced from the array
2. second (spent) input, previously index 1, moved to index 0
3. utxo at index 1 (valid) would be spliced

The give that followed would be using a spent utxo in its inputs,
resulting in the `txn-mempool-conflict`.

This commit adjusts the reconcile logic so that only valid utxos are
kept in a new array, and this array is set as the WalletKey's utxo set
after reconciliation completes.