adjust amount check conditional (#35)
Behavior was observed where a user attempting to withdraw their entire
wallet balance would cause an error to be thrown. This error was caused
by a conditional that was not allowing the tx to be crafted if the
withdraw amount equaled their total balance. This commit fixes the
conditional to ensure the tx gets created before broadcasting.