Cleanup and optimizations (#36)
* platforms: move `on` handlers to Platform interface

There's no reason to maintain separate, unused declarations on the
individual platform modules. Moving these to the Platform interface
allows LotusBot module to appropriately type the handlers and is simpler
to maintain.

* name specific platforms

This commit hard-codes the available platforms in LotusBot. This is to
prepare for type consistency throughout the code, instead of dynamically
accessing collections and getting the `any` type as a result.

* update/use Platform types + update imports

This commit renames the Platform types. It also updates the import in
LotusBot to import all exported data individually instead of importing
all as "Platforms".

* LotusBot: add `_checkAccountValid` method

This method consolidates the logic of checking for and saving/loading
a user's account to/from the database into a single method.

* generate user `secret` in LotusBot

Previous behavior was importing and using the `Util.newUUID()` method
directly in the database. This commit generates the secret in LotusBot
and then passes the data to the `Database.saveAccount()` method instead.

* make `AccountKey` type tuple vs array

* fix `on` handler `message` type

Was still set to `DiscordMessage` after copy-pasta. This commit fixes
that. :)

* prisma: add some table indexes

* wallet: optimize `_reconcileUtxos`

* remove `toLocaleXPI`

Don't localize the XPI amount. This has caused some UX issues with users
that attempt to withdraw their full balance by doing copypasta with the
balance provided by the bot.

* remove unneeded methods, add LotusBot comments

* telegram: set up group message handler

Previous configuration did not have a group message handler. This commit
sets up that handler and moves the "give" command processing to this
handler

* wallet: small cleanup

* discord: remove `_handleReady` log message