Wallet optimization (#38)
* index userId-to-accountId mappings

For #33, logic was implemented that structured `AccountKey`s on-the-fly
from the in-memory `WalletKey`s (i.e. `_getKeys` method). For wallet
operations, processing each `WalletKey` could be very time-consuming in
the case where there are many `WalletKey`s loaded into memory.

This commit adds an `accounts` index, set as an object with `accountId`
as the keys and an array of `userId` strings as the values. This makes
getting `WalletKey` data of an account much faster.

* Remove `accountId` from `WalletKey` type

With the previous commit, this is no longer needed.