General cleanup (#41)
* update README.md

* update install.sh

* README.md: add note about `.env` for auto install

* remove `PlatformDatabaseTable` type

Not using this type anywhere except in the database module, and the
database already implements that type manually for platform-specific
database queries.

* telegram: adjust `withdraw` error responses

* LotusBot: optimize `_saveDeposit`

Previous behavior required a second database query after saving the
deposit in order to fetch the relevant `accountId`. This commit changes
that logic to return the relevant `accountId` when creating the deposit.

Some of the for loop logic is also adjusted to be a bit cleaner.

* LotusBot: logs and errors

Some `_log` messages had duplicate data (e.g. txid), and some try/catch
blocks needed to be updated with the correct method names.

* .env.example: add note on disabling platforms

* LotusBot: try/catch all platform replies

The `balance` and `deposit` handlers were missing the try/catch blocks
for platform replies. This commit fixes that.

* gracefully shutdown on fatal error

Fatal errors for command handlers were not being handled properly. This
commit ensures that fatal error messages are logged and the submodules
are properly shut down.

* _saveDeposit: minor logic cleanup

* _checkAccountValid: add try/catch block

* getAccountBalance: add try/catch block