Subtract fees from give/withdraw amount
Behavior was observed where a withdrawal would fail due to lack of
adequate fees. This commit fixes that issue by ensuring that adequate
fees are provided by subtracting the fees from the output amount. This
is required in the event that the amount in the gathered UTXOs is less
than the output amount + fees.