Implement `backup` command (#43)
* Implement `backup` command

This commit adds the `backup` command for platforms Telegram and
Discord. `backup` allows platform users to dump their wallet seed phrase
associated with the platform account for use in external wallets (e.g.
sendlotus.com).

* Add Discord command registration

Forgot to add the `backup` command registration; this commit fixes that.

* Check account is valid when processing command

Forgot to add the call to `_checkAccountValid` 🙃 this commit fixes
that.

* Get mnemonic from database by `userId`

After adding the call to `_checkAccountValid`, we can use the returned
`userId` to fetch the mnemonic without needing to traverse through an
extra FOREIGN KEY in the database. This commit adjusts the database
query accordingly. Small improvement :)