Fix issue 48 (#50)
* WalletManager: Add Withdraw/Give change outIdx static props

We hard-code the outIdx for the Withdraw and Give transactions as static
properties for use elsewhere in other code that may require them.

* LotusBot: _saveDeposit: Adjust withdrawl conditional

We accept a Withdrawal as a Deposit if the outIdx of the detected utxo
does NOT match the WalletManager.WITHDRAW_CHANGE_OUTIDX or
WalletManager.GIVE_CHANGE_OUTIDX. This change is required in cases where
a user withdraws XPI to another Mrs. Turtle wallet.