General fixes (#55)
* discord.ts: Withdraw error as ephemeral

Previous behavior would send a public withdrawal failure response. This
commit makes the error response ephemeral (i.e. private)

* Fix thrown error strings

After changing some method names, the thrown error strings were not
updated with the new names. This commit fixes that.

* Change wallet tx config object structure

This commit makes the `tx` property an oject with its own properties to
be expanded later if necessary.

* lotusbot.ts: ignore uninitialized bots on shutdown

Previous behavior would occasionally cause uncaught exceptions when
a bot would fail to launch. This commit ensures that only instantiated
bots are shut down in the event of an exception or a graceful shutdown.